Testcase 도메인 객체와 엔티티 객체 분리
