+++ code-challenge/src/main/java/com/megamaker/codechallenge/problem/testcase/domain/TestcaseRepository.java
@@ -1,0 +1,9 @@
+package com.megamaker.codechallenge.problem.testcase.domain;
+
+import com.megamaker.codechallenge.problem.domain.Problem;
+
+import java.util.List;
+
+public interface TestcaseRepository {
+    List<Testcase> findByProblemId(Problem problem);
+}
