+++ code-challenge/src/main/java/com/megamaker/codechallenge/problem/testcase/infra/TestcaseJpaRepository.java
@@ -1,0 +1,10 @@
+package com.megamaker.codechallenge.problem.testcase.infra;
+
+import com.megamaker.codechallenge.problem.infra.ProblemEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface TestcaseJpaRepository extends JpaRepository<TestcaseEntity, Long> {
+    List<TestcaseEntity> findByProblemEntity(ProblemEntity problemEntity);
+}
