--- code-challenge/src/main/java/com/megamaker/codechallenge/testcase/domain/TestcaseRepository.java
@@ -1,7 +1,0 @@
-package com.megamaker.codechallenge.testcase.domain;
-
-import java.util.List;
-
-public interface TestcaseRepository {
-    List<Testcase> findByProblemId(Long problemId);
-}
